forcibly kill only if epmd -kill is not possible

+++ fabfile/tasks/rabbitmq.py
@@ -118,8 +118,9 @@ def allow_rabbitmq_port():
 def stop_rabbitmq_and_set_cookie(uuid):
      with settings(warn_only=True):
          run("service rabbitmq-server stop")
-         run("pkill beam.smp")
-         run("pkill epmd")
+         if 'Killed' not in run("epmd -kill"):
+             run("pkill beam.smp")
+             run("pkill epmd")
          run("rm -rf /var/lib/rabbitmq/mnesia/")
      run("echo '%s' > /var/lib/rabbitmq/.erlang.cookie" % uuid)
      run("chmod 400 /var/lib/rabbitmq/.erlang.cookie")
